Refactor CrSql service and rename schema module
- Load CR-SQLite extension automatically when service starts -
Consolidate operations into single CrSql service (drop CrSql.CrSql.*) -
Add automatic finalizer registration for crsql_finalize() - Rename
schema.ts to CrSqlSchema.ts and re-export as CrSqlSchema - Update
Effect.fn identifiers and example docs - Remove CrSqlClient re-export
and adjust Sqlite client usage
